replaced to constants
